working js without csvs, still can't confirm its runs in browser because of npm start running old code
