+++ index.html
@@ -13,7 +13,7 @@
     <div class="disclaimer">
         Disclaimer: This program is inspired by, but has no association with, the Australian Broadcast Corporation national radio station <a href="https://www.abc.net.au/triplej">Triple J</a>.
     </div>
-    <script src="index.js"></script>
+    <script src="bundle.js" defer></script>
 </body>
 
 </html>
