--- index.ts
@@ -1,433 +1,0 @@
-import * as Papa from 'papaparse';
-interface Playlist {
-    track: string[];
-    artist: string[];
-    weight: number[];
-    [key: string]: any[];
-}
-
-const bee: Playlist = {
-    track: [
-        "Late Night",
-        "Because You Move Me",
-        "Hurricane",
-        "Yesteryear",
-        "I Don't Wanna Talk (I Just Wanna Dance)",
-        "Weekend",
-        "NORTH",
-        "I Luv U Baby (Radio Mix)",
-        "Classic",
-        "I Feel Better",
-        "Late Night Feelings (feat. Lykke Li)",
-        "Used To Love (with Dean Lewis)",
-        "Feed The Fire (feat. Lucky Daye)",
-        "Lucky",
-        "Cry",
-        "Streets - Disclosure Remix",
-        "Lost In Yesterday",
-        "Chemicals (feat. Master Peace) - Remix",
-        "The Prize (feat. Bridgette Amofah)",
-        "Things I Thought Were Mine",
-        "Catch Your Breath",
-        "There For You",
-        "Ticks Me Off",
-        "Think About Things",
-        "Good Times",
-        "Let You Know",
-        "YOU'RE THE ONE",
-        "Stay (feat. Dana Williams)",
-        "Elliot's Song - From \"Euphoria\" An HBO Original Series",
-        "Tondo",
-        "Friends",
-        "Solid Gold",
-        "Lose My Cool - Franc Moody Remix",
-        "Keep Moving - Dave Lee Remix Edit",
-        "deep end",
-        "Good Karma Lover",
-        "Love Again",
-        "Heat Waves",
-        "In My Mind - Northeast Party House Remix",
-        "Can't Do Without You",
-        "R U HIGH (feat. Mallrat)",
-        "Adore You",
-        "FEEL SOMETHING DIFFERENT",
-        "I’m Tired (with Zendaya) - Bonus Track",
-        "C.U.D.I (Can U Dig It)",
-        "Foreign Language",
-        "Caribbean Blue - 2009 Remaster",
-        "All For Us - from the HBO Original Series Euphoria",
-        "Remember",
-        "Feed The Rhythm",
-        "INDUSTRY BABY (feat. Jack Harlow)",
-        "Still Don't Know My Name",
-        "Is It True",
-        "Butterflies (with Starrah & Four Tet)",
-        "Waves",
-        "Blue - Flume Remix",
-        "Damage",
-        "Another Chance",
-        "It's Not The Same Anymore",
-        "Numb",
-        "Jet Black (feat. Brandy)",
-        "Sundress",
-        "Loving Is Easy",
-        "Have It All (feat. Age.Sex.Location)",
-        "When I R.I.P.",
-        "Cold Heart - PNAU Remix",
-        "The Deepest Sighs, the Frankest Shadows",
-        "Morning In Love - Original Mix",
-        "Rushing Back",
-        "Freed from Desire",
-        "Juice",
-        "When I'm Around You",
-        "Straight Up Relaxin' (feat. Cosmo's Midnight)",
-        "This Thing We Got",
-        "good 4 u",
-        "STAY (with Justin Bieber)",
-        "Waiting for Nothing (feat. Yaeger)",
-        "DYSFUNCTIONAL",
-        "Photo ID (with Dominic Fike)",
-        "Where Are You Now",
-        "Crossing Borders (feat. Fritz Kalkbrenner) [Kolombo Remix]",
-        "Dribble",
-        "Feel My Needs",
-        "Live Without Your Love (with Steve Lacy)",
-        "Sorry",
-        "Love Signs",
-        "Let's Go (Remix) (feat. GoldLink)",
-        "Lavender",
-        "Don't Start Now",
-        "Happy People",
-        "LMK",
-        "Everybody Dance",
-        "Borderline",
-        "In My Arms",
-        "Mighty Knight",
-        "Lights Up (feat. Channel Tres)",
-        "Day 'N' Nite (Nightmare)",
-        "Music Sounds Better with You",
-        "Your Song",
-        "Like a Prayer"
-    ],
-    artist: [
-        "Tobiahs",
-        "Tinlicker, Helsloot",
-        "Kanye West",
-        "Cosmo's Midnight",
-        "Glass Animals",
-        "Flume, Moses Sumney",
-        "The Kite String Tangle",
-        "The Original",
-        "F-POS, Gold Fang, Moody Beach",
-        "Hot Chip",
-        "Mark Ronson, Lykke Li",
-        "Martin Garrix, Dean Lewis",
-        "SG Lewis, Lucky Daye",
-        "PNAU, Vlossom",
-        "Cat & Calmell",
-        "Doja Cat, Disclosure",
-        "Tame Impala",
-        "SG Lewis, Master Peace",
-        "The Kite String Tangle, Bridgette Amofah",
-        "Alfie Templeman",
-        "Vlossom",
-        "Gorgon City, MK",
-        "Lime Cordiale",
-        "Daði Freyr",
-        "Genesis Owusu",
-        "Flume, London Grammar",
-        "KAYTRANADA, Syd",
-        "The Aston Shuffle, Fabich, Dana Williams",
-        "Dominic Fike, Zendaya",
-        "Disclosure, Eko Roosevelt",
-        "Flume, Reo Cragun",
-        "PNAU, Kira Divine, Marques Toliver",
-        "Amber Mark, Franc Moody",
-        "Jungle, Dave Lee",
-        "Lykke Li",
-        "Vlossom",
-        "Dua Lipa",
-        "Glass Animals",
-        "Alice Ivy, Ecca Vandal, Northeast Party House",
-        "Caribou",
-        "The Knocks, Mallrat",
-        "Harry Styles",
-        "Bea Miller, Aminé",
-        "Labrinth, Zendaya",
-        "Cosmo's Midnight",
-        "Flight Facilities, Jess",
-        "Enya",
-        "Labrinth, Zendaya",
-        "Becky Hill, David Guetta",
-        "Client Liaison",
-        "Lil Nas X, Jack Harlow",
-        "Labrinth",
-        "Tame Impala",
-        "Skrillex, Starrah, Four Tet",
-        "Kanye West",
-        "Eiffel 65, Flume",
-        "PARTYNEXTDOOR, Halsey",
-        "Roger Sanchez",
-        "Rex Orange County",
-        "Elderbrook",
-        "Anderson .Paak, Brandy",
-        "A$AP Rocky",
-        "Rex Orange County, Benny Sings",
-        "Cosmo's Midnight, Age.Sex.Location",
-        "Labrinth",
-        "Elton John, Dua Lipa, PNAU",
-        "Gang of Youths",
-        "Addal, Lisa May",
-        "Flume, Vera Blue",
-        "Gala",
-        "Young Franco, Pell",
-        "Running Touch",
-        "Yung Bae, Cosmo's Midnight",
-        "The Kite String Tangle",
-        "Olivia Rodrigo",
-        "The Kid LAROI, Justin Bieber",
-        "Hayden James, Yaeger",
-        "KAYTRANADA, VanJess",
-        "Remi Wolf, Dominic Fike",
-        "Lost Frequencies, Calum Scott",
-        "Booka Shade",
-        "Sycco",
-        "WEISS",
-        "Love Regenerator, Steve Lacy, Calvin Harris",
-        "Madonna",
-        "The Jungle Giants",
-        "Khalid, GoldLink",
-        "Disclosure, Channel Tres",
-        "Dua Lipa",
-        "Wh0, Byron Stingily",
-        "Kelela",
-        "Nile Rodgers, Cedric Gervais, Franklin",
-        "Tame Impala",
-        "Purple Disco Machine",
-        "Dubbel Drop",
-        "Flight Facilities, Channel Tres",
-        "Kid Cudi",
-        "NEIL FRANCES",
-        "Elton John",
-        "Madonna"
-    ],
-    weight: [
-        110, 109, 108, 107, 106, 105, 104, 103, 102, 101, 90, 89, 88, 87, 86, 85, 84, 83, 82, 81, 80, 79, 78, 77, 76, 75, 74, 73, 72, 71, 70, 69, 68, 67, 66, 65, 64, 63, 62, 61, 60, 59, 58, 57, 56, 55, 54, 53, 52, 51, 50, 49, 48, 47, 46, 45, 44, 43, 42, 41, 40, 39, 38, 37, 36, 35, 34, 33, 32, 31, 30, 29, 28, 27, 26, 25, 24, 23, 22, 21, 20, 19, 18, 17, 16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1
-    ]
-};
-
-const gaz: Playlist = {
-    track: [
-        "Through The Wire", "Family Business", "The Downeaster 'Alexa'", "Jesus Walks", "Jail",
-        "Say Nothing (feat. MAY-A)", "brothers", "Waiting For Nothing (feat. Yaeger)", "Never Let Me Down",
-        "Remember", "Where Are Ü Now (with Justin Bieber)", "Friends", "Hey Mama", "Sunflower - Spider-Man: Into the Spider-Verse",
-        "What Do You Mean?", "All For Us - from the HBO Original Series Euphoria", "Waves", "Blue - Flume Remix",
-        "One Right Now (with The Weeknd)", "Elliot's Song - From \"Euphoria\" An HBO Original Series", "Hurricane",
-        "Let's Go (Remix) (feat. GoldLink)", "May I Have This Dance (Remix) [feat. Chance the Rapper]", "Ghost",
-        "Without You (feat. Sandro Cavazza)", "Kill For Your Love", "Your Song", "Sacrifice", "Holy (feat. Chance The Rapper)",
-        "Hold My Girl", "FEEL SOMETHING DIFFERENT", "Home", "Conversations with my Wife", "When You Love Someone",
-        "Excavate (feat. Saint Claire)", "Cold Heart - PNAU Remix", "CHANCES", "Get It Right (feat. MØ & GoldLink) - Remix",
-        "Good Old Days (feat. Kesha)", "Unholy (feat. Kim Petras)", "Happier", "THATS WHAT I WANT", "You’re The Best Thing About Me - U2 Vs. Kygo",
-        "There for You", "Good Grief - Don Diablo Remix", "Lemon", "I Think Of You", "Promises (with Sam Smith)", "Late Night",
-        "These Days (feat. Jess Glynne, Macklemore & Dan Caplen)", "Jet Black (feat. Brandy)", "Pussy Is God", "Beautiful People (feat. Khalid)",
-        "Carry Me (feat. Julia Michaels)", "Anyone", "Mighty Knight", "Love Signs", "First Time", "Shots - Broiler Remix",
-        "I Don't Care (with Justin Bieber)", "Good Times", "Paradise", "Ballin Flossin", "Sundress", "One Day (feat. Ryan Tedder)",
-        "Coke & Henny Pt. 2", "Blinding Lights", "In My Mind", "Angels (feat. Saba)", "Tints (feat. Kendrick Lamar)", "Is This Love - Remix",
-        "Rushing Back", "INDUSTRY BABY (feat. Jack Harlow)", "I’m Tired (with Zendaya) - Bonus Track", "STAY (with Justin Bieber)",
-        "Weekend", "September - Throttle x Earth, Wind & Fire", "Lift Off", "Surfin'", "3005", "Giant (with Rag'n'Bone Man)",
-        "With You", "Like Gold", "I Fall Apart", "Heatstroke (feat. Young Thug, Pharrell Williams & Ariana Grande)", "Stranger Things (feat. OneRepublic)",
-        "Fast Car", "Another Chance", "Stay With Me (with Justin Timberlake, Halsey, & Pharrell)", "Yummy", "How to Play the Flute (feat. King Draino)",
-        "Didn't I", "Feels (feat. Pharrell Williams, Katy Perry & Big Sean)", "My People", "Ball For Me (feat. Nicki Minaj)", "Cross Me (feat. Chance the Rapper & PnB Rock)",
-        "Audio (feat. Sia, Diplo, and Labrinth)", "Be As You Are - JordanXL Remix", "Young Dumb & Broke", "Let You Know"
-    ],
-    artist: [
-        "Kanye West", "Kanye West", "Billy Joel", "Kanye West", "Kanye West", "Flume, MAY-A", "Gang of Youths", "Hayden James, Yaeger",
-        "Kanye West, JAY-Z, J. Ivy", "Becky Hill, David Guetta", "Jack Ü, Skrillex, Diplo, Justin Bieber", "Flume, Reo Cragun", "Kanye West",
-        "Post Malone, Swae Lee", "Justin Bieber", "Labrinth, Zendaya", "Kanye West", "Eiffel 65, Flume", "Post Malone, The Weeknd",
-        "Dominic Fike, Zendaya", "Kanye West", "Khalid, GoldLink", "Francis and the Lights, Chance the Rapper", "Justin Bieber",
-        "Avicii, Sandro Cavazza", "Labrinth", "Elton John", "The Weeknd", "Justin Bieber, Chance the Rapper", "George Ezra", "Bea Miller, Aminé",
-        "Vince Staples, Richie Kohan", "Jon Bellion", "James TW", "Macklemore, Saint Claire", "Elton John, Dua Lipa, PNAU", "KAYTRANADA, Shay Lia",
-        "Diplo, MØ, GoldLink", "Macklemore, Kesha", "Sam Smith, Kim Petras", "Marshmello, Bastille", "Lil Nas X", "U2, Kygo", "Martin Garrix, Troye Sivan",
-        "Bastille, Don Diablo", "N.E.R.D, Rihanna, Pharrell Williams", "Jeremih, Chris Brown, Big Sean", "Calvin Harris, Sam Smith, Jessie Reyez",
-        "Tobiahs", "Rudimental, Jess Glynne, Macklemore, Dan Caplen", "Anderson .Paak, Brandy", "King Princess", "Ed Sheeran, Khalid", "Kygo, Julia Michaels",
-        "Justin Bieber", "Dubbel Drop", "The Jungle Giants", "Kygo, Ellie Goulding", "Imagine Dragons, Broiler", "Ed Sheeran, Justin Bieber", "Genesis Owusu",
-        "George Ezra", "Chance the Rapper, Shawn Mendes", "A$AP Rocky", "Logic, Ryan Tedder", "Pink Sweat$", "The Weeknd", "Dynoro, Gigi D'Agostino",
-        "Chance the Rapper, Saba", "Anderson .Paak, Kendrick Lamar", "Bob Marley & The Wailers, LVNDSCAPE, Bolier", "Flume, Vera Blue", "Lil Nas X, Jack Harlow",
-        "Labrinth, Zendaya", "The Kid LAROI, Justin Bieber", "Flume, Moses Sumney", "Throttle, Earth, Wind & Fire", "Labrinth", "Kid Cudi, Pharrell Williams",
-        "Childish Gambino", "Calvin Harris, Rag'n'Bone Man", "Fetty Wap, KDL", "Vance Joy", "Post Malone", "Calvin Harris, Young Thug, Pharrell Williams, Ariana Grande",
-        "Kygo, OneRepublic", "Jonas Blue, Dakota", "Roger Sanchez", "Calvin Harris, Justin Timberlake, Halsey, Pharrell Williams", "Justin Bieber",
-        "Macklemore, King Draino", "OneRepublic", "Calvin Harris, Pharrell Williams, Katy Perry, Big Sean", "REMI, Sensible J", "Post Malone, Nicki Minaj",
-        "Ed Sheeran, Chance the Rapper, PnB Rock", "Sia, Diplo, Labrinth, LSD", "Mike Posner, JordanXL", "Khalid", "Flume, London Grammar"
-    ],
-    weight: [
-        110, 109, 108, 107, 106, 105, 104, 103, 102, 101, 90, 89, 88, 87, 86, 85, 84, 83, 82, 81, 80, 79, 78, 77, 76, 75, 74, 73, 72, 71, 70, 69, 68, 67, 66, 65, 64, 63, 62, 61, 60, 59, 58, 57, 56, 55, 54, 53, 52, 51, 50, 49, 48, 47, 46, 45, 44, 43, 42, 41, 40, 39, 38, 37, 36, 35, 34, 33, 32, 31, 30, 29, 28, 27, 26, 25, 24, 23, 22, 21, 20, 19, 18, 17, 16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1
-    ]
-};
-
-// function csvToPlaylist(csvData: string): Playlist {
-//     const parsedData = Papa.parse<Record<string, string | number>>(csvData, {
-//         header: true,
-//         dynamicTyping: true,
-//         skipEmptyLines: true,
-//     });
-
-//     const playlist: Playlist = {
-//         track: [],
-//         artist: [],
-//         weight: [],
-//     };
-
-//     if (parsedData.errors.length > 0) {
-//         console.error('Error parsing CSV:', parsedData.errors);
-//         return playlist;
-//     }
-
-//     parsedData.data.forEach(row => {
-//         playlist.track.push(row.track as string);
-//         playlist.artist.push(row.artist as string);
-//         playlist.weight.push(row.weight as number);
-
-//         // Handling additional columns dynamically
-//         Object.keys(row).forEach(key => {
-//             if (!['track', 'artist', 'weight'].includes(key)) {
-//                 if (!playlist[key]) {
-//                     playlist[key] = [];
-//                 }
-//                 playlist[key].push(row[key]);
-//             }
-//         });
-//     });
-
-//     return playlist;
-// }
-
-// function getPlaylistNameFromPath(filePath: string): string {
-//     const fileName = filePath.split('/').pop() || '';
-//     return fileName.split('.').shift() || 'UnnamedPlaylist';
-// }
-
-interface CombinedPlaylist {
-    track: string;
-    artist: string;
-    weight: number;
-    sources: { [key: string]: number | null };
-}
-
-function mergePlaylists(playlists: { [key: string]: Playlist }): CombinedPlaylist[] {
-    const combinedList: CombinedPlaylist[] = [];
-
-    Object.keys(playlists).forEach(playlistName => {
-        const playlist = playlists[playlistName];
-
-        playlist.track.forEach((track, index) => {
-            const artist = playlist.artist[index];
-            const weight = playlist.weight[index];
-
-            // Check if the track-artist combination already exists in the combined list
-            const existingEntry = combinedList.find(
-                entry => entry.track === track && entry.artist === artist
-            );
-
-            if (existingEntry) {
-                // If it exists, update the weight and the source index
-                existingEntry.weight += weight;
-                existingEntry.sources[playlistName] = index + 1; // +1 to adjust for 1-based index
-            } else {
-                // If it doesn't exist, create a new entry with nulls for all sources except the current one
-                const newEntry: CombinedPlaylist = {
-                    track,
-                    artist,
-                    weight,
-                    sources: Object.keys(playlists).reduce((acc, key) => {
-                        acc[key] = key === playlistName ? index + 1 : null;
-                        return acc;
-                    }, {} as { [key: string]: number | null }),
-                };
-                combinedList.push(newEntry);
-            }
-        });
-    });
-
-    // Sort the combined list in descending order by weight
-    combinedList.sort((a, b) => b.weight - a.weight);
-
-    return combinedList;
-}
-
-function createCountdown(combinedList: CombinedPlaylist[], countdownNumber: number): CombinedPlaylist[] {
-    // Trim the list to only include the top `countdownNumber` songs
-    const trimmedList = combinedList.slice(0, countdownNumber);
-
-    // Reverse the order of the tracks
-    const reversedList = trimmedList.reverse();
-
-    return reversedList;
-}
-
-// function generatePlaylist() {
-//     alert("hi");
-// }
-
-
-// Ensure the function is accessible globally
-// window.generatePlaylist = generatePlaylist;
-
-function generatePlaylist() {
-    const playlists: { [key: string]: Playlist } = {};
-
-    const COUNTDOWN_NUMBER = 50;
-
-    playlists["bee"] = bee;
-    playlists["gaz"] = gaz;
-
-    console.log(playlists);
-
-    const combinedList = mergePlaylists(playlists);
-    const countdownList = createCountdown(combinedList, COUNTDOWN_NUMBER);
-    console.log(countdownList);
-    // alert(countdownList[COUNTDOWN_NUMBER - 1].track)
-}
-generatePlaylist();
-
-// function generatePlaylist() {
-//     const filePaths = [
-//         'data/website-data/ash.csv',
-//         'data/website-data/bee.csv',
-//         'data/website-data/billy.csv',
-//         'data/website-data/bryce.csv',
-//         'data/website-data/cam.csv',
-//         'data/website-data/eric.csv',
-//         'data/website-data/eve.csv',
-//         'data/website-data/gaz.csv',
-//         'data/website-data/han.csv',
-//         'data/website-data/jack.csv',
-//         'data/website-data/james.csv',
-//         'data/website-data/jim.csv',
-//         'data/website-data/jo.csv',
-//         'data/website-data/josh.csv',
-//         'data/website-data/kara.csv',
-//         'data/website-data/liana.csv',
-//         'data/website-data/loz.csv',
-//         'data/website-data/oscar.csv',
-//         'data/website-data/sam.csv',
-//         'data/website-data/sammy.csv'
-//     ];
-
-//     const playlists: { [key: string]: Playlist } = {};
-
-//     const COUNTDOWN_NUMBER = 50;
-
-//     filePaths.forEach(filePath => {
-//         const csvData = fs.readFileSync(filePath, 'utf8');
-//         const playlistName = getPlaylistNameFromPath(filePath);
-//         playlists[playlistName] = csvToPlaylist(csvData);
-//     });
-
-//     console.log(playlists);
-
-//     const combinedList = mergePlaylists(playlists);
-//     const countdownList = createCountdown(combinedList, COUNTDOWN_NUMBER);
-//     console.log(countdownList);
-//     alert(countdownList[COUNTDOWN_NUMBER - 1].track)
-// }
-// generatePlaylist();
