+++ src/index.ts
@@ -1,50 +1,128 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-function mergePlaylists(playlists) {
-    var combinedList = [];
-    Object.keys(playlists).forEach(function (playlistName) {
-        var playlist = playlists[playlistName];
-        playlist.track.forEach(function (track, index) {
-            var artist = playlist.artist[index];
-            var weight = playlist.weight[index];
+import * as Papa from 'papaparse';
+interface Playlist {
+    track: string[];
+    artist: string[];
+    weight: number[];
+    [key: string]: any[];
+}
+
+// function csvToPlaylist(csvData: string): Playlist {
+//     const parsedData = Papa.parse<Record<string, string | number>>(csvData, {
+//         header: true,
+//         dynamicTyping: true,
+//         skipEmptyLines: true,
+//     });
+
+//     const playlist: Playlist = {
+//         track: [],
+//         artist: [],
+//         weight: [],
+//     };
+
+//     if (parsedData.errors.length > 0) {
+//         console.error('Error parsing CSV:', parsedData.errors);
+//         return playlist;
+//     }
+
+//     parsedData.data.forEach(row => {
+//         playlist.track.push(row.track as string);
+//         playlist.artist.push(row.artist as string);
+//         playlist.weight.push(row.weight as number);
+
+//         // Handling additional columns dynamically
+//         Object.keys(row).forEach(key => {
+//             if (!['track', 'artist', 'weight'].includes(key)) {
+//                 if (!playlist[key]) {
+//                     playlist[key] = [];
+//                 }
+//                 playlist[key].push(row[key]);
+//             }
+//         });
+//     });
+
+//     return playlist;
+// }
+
+// function getPlaylistNameFromPath(filePath: string): string {
+//     const fileName = filePath.split('/').pop() || '';
+//     return fileName.split('.').shift() || 'UnnamedPlaylist';
+// }
+
+interface CombinedPlaylist {
+    track: string;
+    artist: string;
+    weight: number;
+    sources: { [key: string]: number | null };
+}
+
+function mergePlaylists(playlists: { [key: string]: Playlist }): CombinedPlaylist[] {
+    const combinedList: CombinedPlaylist[] = [];
+
+    Object.keys(playlists).forEach(playlistName => {
+        const playlist = playlists[playlistName];
+
+        playlist.track.forEach((track, index) => {
+            const artist = playlist.artist[index];
+            const weight = playlist.weight[index];
+
             // Check if the track-artist combination already exists in the combined list
-            var existingEntry = combinedList.find(function (entry) { return entry.track === track && entry.artist === artist; });
+            const existingEntry = combinedList.find(
+                entry => entry.track === track && entry.artist === artist
+            );
+
             if (existingEntry) {
                 // If it exists, update the weight and the source index
                 existingEntry.weight += weight;
                 existingEntry.sources[playlistName] = index + 1; // +1 to adjust for 1-based index
-            }
-            else {
+            } else {
                 // If it doesn't exist, create a new entry with nulls for all sources except the current one
-                var newEntry = {
-                    track: track,
-                    artist: artist,
-                    weight: weight,
-                    sources: Object.keys(playlists).reduce(function (acc, key) {
+                const newEntry: CombinedPlaylist = {
+                    track,
+                    artist,
+                    weight,
+                    sources: Object.keys(playlists).reduce((acc, key) => {
                         acc[key] = key === playlistName ? index + 1 : null;
                         return acc;
-                    }, {}),
+                    }, {} as { [key: string]: number | null }),
                 };
                 combinedList.push(newEntry);
             }
         });
     });
+
     // Sort the combined list in descending order by weight
-    combinedList.sort(function (a, b) { return b.weight - a.weight; });
+    combinedList.sort((a, b) => b.weight - a.weight);
+
     return combinedList;
 }
-function createCountdown(combinedList, countdownNumber) {
+
+function createCountdown(combinedList: CombinedPlaylist[], countdownNumber: number): CombinedPlaylist[] {
     // Trim the list to only include the top `countdownNumber` songs
-    var trimmedList = combinedList.slice(0, countdownNumber);
+    const trimmedList = combinedList.slice(0, countdownNumber);
+
     // Reverse the order of the tracks
-    var reversedList = trimmedList.reverse();
+    const reversedList = trimmedList.reverse();
+
     return reversedList;
 }
+
+function generatePlaylist() {
+    console.log("hi");
+}
+
+declare global {
+    interface Window {
+        generatePlaylist: () => void;
+        // generatePlaylist2: () => void;
+    }
+}
+
 // Ensure the function is accessible globally
 window.generatePlaylist = generatePlaylist;
 // window.generatePlaylist2 = generatePlaylist2;
-function generatePlaylist() {
-    var bee = {
+
+function generatePlaylist2() {
+    const bee: Playlist = {
         track: [
             "Late Night",
             "Because You Move Me",
@@ -253,7 +331,8 @@ function generatePlaylist() {
             110, 109, 108, 107, 106, 105, 104, 103, 102, 101, 90, 89, 88, 87, 86, 85, 84, 83, 82, 81, 80, 79, 78, 77, 76, 75, 74, 73, 72, 71, 70, 69, 68, 67, 66, 65, 64, 63, 62, 61, 60, 59, 58, 57, 56, 55, 54, 53, 52, 51, 50, 49, 48, 47, 46, 45, 44, 43, 42, 41, 40, 39, 38, 37, 36, 35, 34, 33, 32, 31, 30, 29, 28, 27, 26, 25, 24, 23, 22, 21, 20, 19, 18, 17, 16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1
         ]
     };
-    var gaz = {
+    
+    const gaz: Playlist = {
         track: [
             "Through The Wire", "Family Business", "The Downeaster 'Alexa'", "Jesus Walks", "Jail",
             "Say Nothing (feat. MAY-A)", "brothers", "Waiting For Nothing (feat. Yaeger)", "Never Let Me Down",
@@ -300,16 +379,22 @@ function generatePlaylist() {
             110, 109, 108, 107, 106, 105, 104, 103, 102, 101, 90, 89, 88, 87, 86, 85, 84, 83, 82, 81, 80, 79, 78, 77, 76, 75, 74, 73, 72, 71, 70, 69, 68, 67, 66, 65, 64, 63, 62, 61, 60, 59, 58, 57, 56, 55, 54, 53, 52, 51, 50, 49, 48, 47, 46, 45, 44, 43, 42, 41, 40, 39, 38, 37, 36, 35, 34, 33, 32, 31, 30, 29, 28, 27, 26, 25, 24, 23, 22, 21, 20, 19, 18, 17, 16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1
         ]
     };
-    var playlists = {};
-    var COUNTDOWN_NUMBER = 50;
+    
+    const playlists: { [key: string]: Playlist } = {};
+
+    const COUNTDOWN_NUMBER = 50;
+
     playlists["bee"] = bee;
     playlists["gaz"] = gaz;
+
     console.log(playlists);
-    var combinedList = mergePlaylists(playlists);
-    var countdownList = createCountdown(combinedList, COUNTDOWN_NUMBER);
+
+    const combinedList = mergePlaylists(playlists);
+    const countdownList = createCountdown(combinedList, COUNTDOWN_NUMBER);
     console.log(countdownList);
     // alert(countdownList[COUNTDOWN_NUMBER - 1].track)
 }
+
 // function generatePlaylist() {
 //     const filePaths = [
 //         'data/website-data/ash.csv',
@@ -333,14 +418,19 @@ function generatePlaylist() {
 //         'data/website-data/sam.csv',
 //         'data/website-data/sammy.csv'
 //     ];
+
 //     const playlists: { [key: string]: Playlist } = {};
+
 //     const COUNTDOWN_NUMBER = 50;
+
 //     filePaths.forEach(filePath => {
 //         const csvData = fs.readFileSync(filePath, 'utf8');
 //         const playlistName = getPlaylistNameFromPath(filePath);
 //         playlists[playlistName] = csvToPlaylist(csvData);
 //     });
+
 //     console.log(playlists);
+
 //     const combinedList = mergePlaylists(playlists);
 //     const countdownList = createCountdown(combinedList, COUNTDOWN_NUMBER);
 //     console.log(countdownList);
